Add user tabs to profile page
This commit adds two new tabs, "View" and "Edit", to the user profile page. The tabs are implemented using the TabBar component from @silintl/ui-components. The UserTabs.svelte file was added to implement this feature.

The global.css file was also updated with a new font-family and line-height property, and the primary color was changed.

Finally, some code cleanup was done in _layout.svelte by removing an empty object from menuItems.